libstdgit: introduce Git Standard Library
The Git Standard Library intends to serve as the foundational library
and root dependency that other libraries in Git will be built off
of. That is to say, suppose we have libraries X and Y; a user that
wants to use X and Y would need to include X, Y, and this Git Standard
Library.

As the first library in Git, new Makefile rules are written for
libstdgit.a and for stubbed out files, libstubgit.a.

Unit tests will be added in a separate commit.

Signed-off-by: Calvin Wan <calvinwan@google.com>
Signed-off-by: Phillip Wood <phillip.wood123@gmail.com>
Signed-off-by: Jonathan Tan <jonathantanmy@google.com>